backend/requirements.txt
django==4.2
djangorestframework==3.14.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0
django-phonenumber-field==7.1.0
djoser==2.2.0
phonenumbers==8.13.19
django-filter
gunicorn==21.2.0